+++ src/golang/net/index.ts
@@ -1,2 +1,4 @@
 export { JoinHostPort } from './JoinHostPort.ts'
+export { ParseCIDR } from './ParseCIDR.ts'
+export { ParseIP } from './ParseIP.ts'
 export { SplitHostPort } from './SplitHostPort.ts'