+++ src/golang/subtle/index.ts
@@ -1,1 +1,4 @@
 export { ConstantTimeCompare } from './ConstantTimeCompare.ts'
+export { ConstantTimeCopy } from './ConstantTimeCopy.ts'
+export { ConstantTimeEq } from './ConstantTimeEq.ts'
+export { ConstantTimeSelect } from './ConstantTimeSelect.ts'