+++ src/python/re/index.ts
@@ -1,1 +1,3 @@
+export { findall } from './findall.ts'
+export { finditer } from './finditer.ts'
 export { sub } from './sub.ts'